PhotoShare - Commands modified to use session service

+++ DB Advanced/PhotoShareSystem/PhotoShare.Client/Core/Commands/DeleteUserCommand.cs
@@ -9,10 +9,12 @@
     public class DeleteUserCommand : ICommand
     {
         private readonly IUserService userService;
+        private readonly ISessionService sessionService;
 
-        public DeleteUserCommand(IUserService userService)
+        public DeleteUserCommand(IUserService userService, ISessionService sessionService)
         {
             this.userService = userService;
+            this.sessionService = sessionService;
         }
 
         // DeleteUser <username>
@@ -25,6 +27,13 @@
             if (!userExists)
             {
                 throw new ArgumentException($"User {username} not found!");
+            }
+
+            var userId = this.userService.ByUsername<UserDto>(username).Id;
+
+            if (!this.sessionService.IsLoggedIn(userId))
+            {
+                throw new InvalidOperationException("Invalid credentials!");
             }
 
             this.userService.Delete(username);
